Add Interactable lint test
Since #19174 bounds are calculated from WDist instead of pixels. This causes many of the older maps to crash games. And to prevent crashing, we add a lint test.

As calculating the bounds requires MapGrid, we need to write a separate lint test for Interactable instead of defaulting to RulesetLoaded checks.